Refactor Players
